[fix] 키워드 검색 결과 수정(totalElement)

+++ encyclo-service/src/main/java/com/example/encycloservice/domain/SearchPlantQueryResult.java
@@ -11,7 +11,7 @@ import java.util.List;
 @Builder
 public class SearchPlantQueryResult {
 
-    private int totalElement;
+    private long totalElement;
     private int totalPage;
     private List<SingleSearchPlantResult> results;
